Move helper functions to Pest
